requirements.txt
# ─────────────────────────────────────────────────────────────
# AI Resume–JD Matching & Skill Gap Analyzer — requirements.txt
# ─────────────────────────────────────────────────────────────

# ── Web Framework ──────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# ── NLP & Embeddings ───────────────────────────────────────────
spacy>=3.7.4
sentence-transformers>=2.7.0
transformers>=4.40.2
torch>=2.3.0
scikit-learn>=1.4.2
nltk>=3.8.1

# ── Text Extraction ────────────────────────────────────────────
pdfplumber>=0.11.0
python-docx>=1.1.0
beautifulsoup4>=4.12.3
requests>=2.31.0
lxml>=5.2.1

# ── Database ───────────────────────────────────────────────────
sqlalchemy>=2.0.30
aiosqlite>=0.20.0
alembic>=1.13.1

# ── Data Processing ────────────────────────────────────────────
pandas>=2.2.2
numpy>=1.26.4
pyarrow>=16.0.0
kaggle>=1.6.14

# ── Frontend ───────────────────────────────────────────────────
streamlit>=1.34.0
plotly>=5.21.0
wordcloud>=1.9.3
matplotlib>=3.8.4
Pillow>=10.3.0

# ── LLM / AI ───────────────────────────────────────────────────
openai>=1.30.1
tiktoken>=0.7.0

# ── Config & Utils ─────────────────────────────────────────────
pydantic>=2.7.1
pydantic-settings>=2.2.1
python-dotenv>=1.0.1
structlog>=24.1.0
httpx>=0.27.0
aiofiles>=23.2.1
tenacity>=8.3.0

# ── Testing ────────────────────────────────────────────────────
pytest>=8.2.0
pytest-asyncio>=0.23.6
pytest-cov>=5.0.0
httpx>=0.27.0

# ── Dev Tools ──────────────────────────────────────────────────
black>=24.4.2
isort>=5.13.2
flake8>=7.0.0
